Exception handling

+++ src/main/java/com/projectcommerce/dscommerce/Services/ProductService.java
@@ -1,14 +1,20 @@
-package com.projectcommerce.dscommerce.Services;
+package com.projectcommerce.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.projectcommerce.dscommerce.DTO.ProductDTO;
-import com.projectcommerce.dscommerce.Entities.Product;
-import com.projectcommerce.dscommerce.Repositories.ProductRepository;
+import com.projectcommerce.dscommerce.entities.Product;
+import com.projectcommerce.dscommerce.repositories.ProductRepository;
+import com.projectcommerce.dscommerce.services.exceptions.DatabaseException;
+import com.projectcommerce.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 
@@ -20,7 +26,8 @@ public class ProductService {
     
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).orElseThrow(
+            () -> new ResourceNotFoundException("Resource not found"));
         return new ProductDTO(product);
     }
 
@@ -40,15 +47,30 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {  Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+
+        catch (EntityNotFoundException e)   {
+            throw new ResourceNotFoundException("Resource not found");
+        }  
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try{
+            repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e)    {
+            throw new DatabaseException("Integrity Violation");
+            
+        }
+    
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity){
